// src/modules/storefront/interfaces/Theme.ts
export interface MainModuleThemeInterface {
  name: string;
  type: ModulesType;
  id: string;
  styleData?: any;
  contentData?: any;
  deviceType?: 'none' | 'mobile' | 'desktop';
  languageType?: string;
  mobileStyleData?: any;
  mobileContentData?: any;
}
// ...
export interface ProductPageData extends MainModuleThemeInterface {
  type: ModulesType.PRODUCT_PAGE;
  styleData: {
    margin?: string;
    padding?: string;
    backTextColor?: string;
    backBackgroundColor?: string;
    backgroundColor?: string;
    textColor?: string;
    categoriesTagBackgroundColor?: string;
    categoriesTagTextColor?: string;
    categoriesTextColor?: string;
    descriptionTextColor?: string;
    priceTextColor?: string;
    nameTextColor?: string;
    actionButton?: boolean;
    buttonText?: string;
    buttonTextColor?: string;
    buttonColor?: string;
    showBlockchainInfo?: boolean;
    showValue?: boolean;
    showDescription?: boolean;
    showCategory?: boolean;
    showProductName?: boolean;
    blockchainInfoBackgroundColor?: string;
    blockchainInfoTextColor?: string;
  };
  mobileStyleData: ProductPageData['styleData'];
}
// ...
export enum ModulesType {
  HEADER = 'Header',
  CONFIGURATION = 'Configuration',
  CATEGORIES = 'Categories',
  BANNER = 'Banner',
  CARDS = 'Cards',
  FOOTER = 'Footer',
  COOKIE = 'Cookie',
  ACCORDIONS = 'Accordions',
  IMAGE_PLUS_TEXT = 'Imagem + Texto',
  PARAGRAPH = 'Paragraph',
  PRODUCT_PAGE = 'Product_page',
  MIDIA = 'Midia',
  GRID_ITEM_AREA = 'GridItemArea',
  DYNAMIC_API = 'DynamicApi',
  TABLE = 'Table',
  BANNER_VARIANT = 'BannerVariant',
  CONTENT_CARD = 'Content Card',
  BUTTON = 'Button',
  BANNER_WJJC = 'BannerWJJC',
  PASS_BENEFIT = 'Pass - Benefício',
  REDIRECT = 'Redirect',
}